[feature/#22] 서버통신 완료
