+++ app/src/main/java/com/teamfillin/fillin/data/repository/AuthRepositoryImpl.kt
@@ -2,7 +2,9 @@ package com.teamfillin.fillin.data.repository
 
 import com.teamfillin.fillin.data.local.FillInDataStore
 import com.teamfillin.fillin.data.response.ResponseAuth
+import com.teamfillin.fillin.data.response.toEntity
 import com.teamfillin.fillin.data.service.AuthService
+import com.teamfillin.fillin.domain.entity.Auth
 import com.teamfillin.fillin.domain.repository.AuthRepository
 import retrofit2.await
 import timber.log.Timber
@@ -12,19 +14,16 @@ class AuthRepositoryImpl @Inject constructor(
     private val dataStore: FillInDataStore,
     private val service: AuthService
 ) : AuthRepository {
-    override suspend fun login(token: String): Result<Boolean> {
+    override suspend fun login(token: String) {
         runCatching {
             service.login(token).await()
         }.fold({
+            val auth = it.data.toEntity()
             with(dataStore) {
-                userToken = it.data.accessToken
-                refreshToken = it.data.refreshToken
+                userToken = auth.accessToken
+                refreshToken = auth.refreshToken
             }
-            if (it.data is ResponseAuth.SignUp) dataStore.nickname = it.data.nickName
-            return Result.success(true)
-        }, {
-            Timber.e(it)
-            return Result.failure(it)
-        })
+            if (auth is Auth.SignUp) dataStore.nickname = auth.nickName
+        }, Timber::e)
     }
 }
