+++ app/src/main/java/com/teamfillin/fillin/data/response/ResponseAuth.kt
@@ -1,20 +1,15 @@
 package com.teamfillin.fillin.data.response
 
-sealed class ResponseAuth(
-    open val email: String?,
-    open val accessToken: String,
-    open val refreshToken: String
-) {
-    data class Login(
-        override val email: String?,
-        override val accessToken: String,
-        override val refreshToken: String
-    ) : ResponseAuth(email, accessToken, refreshToken)
+import com.teamfillin.fillin.domain.entity.Auth
 
-    data class SignUp(
-        override val email: String?,
-        val nickName: String,
-        override val accessToken: String,
-        override val refreshToken: String
-    ) : ResponseAuth(email, accessToken, refreshToken)
+data class ResponseAuth(
+    val email: String?,
+    val nickName: String? = null,
+    val accessToken: String,
+    val refreshToken: String
+)
+
+fun ResponseAuth.toEntity(): Auth = when (nickName) {
+    null -> Auth.Login(email, accessToken, refreshToken)
+    else -> Auth.SignUp(email, nickName, accessToken, refreshToken)
 }
