+++ app/src/main/java/com/teamfillin/fillin/domain/repository/AuthRepository.kt
@@ -1,5 +1,5 @@
 package com.teamfillin.fillin.domain.repository
 
 interface AuthRepository {
-    suspend fun login(token: String): Result<Boolean>
+    suspend fun login(token: String)
 }
